Added some animations and controller
-changed it to running in place
