+++ Assets/Animations.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9c84f708001b00f4da89927b795a9f56
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
