+++ Assets/Animations/Idle.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d6f510a16f83eae47ad8cab6336f6d6e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
