+++ Assets/Animations/PlayerController.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5092a43d04753464189b77a9c53111c6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
