+++ Assets/Animations/Running.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 24e6f8ee4fa5140428be9c8842589727
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
