+++ Assets/Animations/Walking.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0e8cace569ed5484095d2076f42c44ee
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
