+++ Assets/Imported Animation Assets.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 77d045f1d059ee14293b0fdbb42b23b2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
