--- Assets/MoveCamera.cs
@@ -1,15 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class MoveCamera : MonoBehaviour
-{
-
-    public Transform cameraPosition;
-   
-    // Update is called once per frame
-    void Update()
-    {
-        transform.position = cameraPosition.position;
-    }
-}
