--- Assets/PlayerCam.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3b164c119caa7e3498ff79b2f420c8a9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
