+++ Assets/Scripts/MoveCamera.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0da7dfdad02b0d940aaf1df4b893636b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
