+++ Assets/Scripts/Player_Movement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aa94d9c059a43f243bc6164f78206e8d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
